+++ bundles/statistics/statsgrid2016/Flyout.js
@@ -163,14 +163,6 @@ Oskari.clazz.define('Oskari.statistics.statsgrid.Flyout',
             });
 
             accordion.insertTo(el);
-            // barchart.createChart(el);
-            // Add grid
-            var grid = Oskari.clazz.create('Oskari.statistics.statsgrid.Datatable', sb, this.instance.getLocalization());
-            grid.showRegionsetSelector(!isEmbedded);
-            grid.showIndicatorRemoval(!isEmbedded);
-            grid.render(el);
-
-            me._grid = grid;
 
         },
         closePanels: function() {