+++ bundles/statistics/statsgrid2016/instance.js
@@ -79,11 +79,8 @@ Oskari.clazz.define(
                 dsiservice.addGroup('indicators', this.getLocalization().dataProviderInfoTitle || 'Indicators');
             }
 
-            // Check if configurated to show indicators on the map for vector format
-            if(conf && conf.vectorViewer === true) {
-                // regionsetViewer creation need be there because of start order
-                this.regionsetViewer = Oskari.clazz.create('Oskari.statistics.statsgrid.RegionsetViewer', this, sandbox, this.conf);
-            }
+            // regionsetViewer creation need be there because of start order
+            this.regionsetViewer = Oskari.clazz.create('Oskari.statistics.statsgrid.RegionsetViewer', this, sandbox, this.conf);
         },
         extendTile: function (el) {
           var container = this.getTile().container.append(el);
@@ -114,7 +111,7 @@ Oskari.clazz.define(
             return jQuery('#contentMap').hasClass('published');
         },
         hasData: function () {
-            return this.statsService.getDatasource().length && this.statsService.getRegionsets().length;
+            return this.statsService.getDatasource().length;
         },
 
         /**
@@ -263,7 +260,10 @@ Oskari.clazz.define(
             },
             FeatureEvent: function(evt) {
                 this.statsService.notifyOskariEvent(evt);
-            }
+            },
+            AfterChangeMapLayerOpacityEvent: function (evt) {
+                 this.statsService.notifyOskariEvent(evt);
+             }
         },
 
         /**